KA-276 Introduce IRetrieveService
IRetrieveService service retrieves items from repository, in my case ToDos. The service is used to get the ToDo, save one ToDo in cache and for checking if the ToDo exists.

+++ ToDoList/src/ToDoList.Api/Controllers/ToDosController.cs
@@ -18,32 +18,45 @@ namespace ToDoList.Api.Controllers
         private readonly IToDoRepository _toDoRepository;
         private readonly IUrlLocationService _locationService;
         private readonly IFormationService _formationService;
+        private readonly IRetrieveService<ToDo> _retrieveService;
 
-        public ToDosController(IToDoRepository toDoRepository, IUrlLocationService locationService, IFormationService formationService)
+        public ToDosController(IToDoRepository toDoRepository, IUrlLocationService locationService, IFormationService formationService, IRetrieveService<ToDo> retrieveService)
         {
+            _retrieveService = retrieveService;
             _formationService = formationService;
             _toDoRepository = toDoRepository;
             _locationService = locationService;
         }
         
         public async Task<IHttpActionResult> GetToDosAsync()
-            => Ok(await _toDoRepository.GetToDosAsync());
+            => Ok(await _retrieveService.RetriveAllAsync());
 
         [Route("{id}", Name = WebApiRoutes.GetToDoRoute)]
         public async Task<IHttpActionResult> GetToDoAsync(Guid id)
-            => Ok(await _toDoRepository.GetToDoAsync(id));
-        
+        {
+            if (!await _retrieveService.IsInDatabaseAsync(id))
+            {
+                return NotFound();
+            }
+
+            return Ok(await _retrieveService.RetriveOneAsync(id));
+        }
+
         public async Task<IHttpActionResult> PostToDoAsync([FromBody]ToDo toDoItem)
         {
-            var newToDo = await _formationService.CreateToDoAsync(toDoItem);
-            var toDoLocationUrl = _locationService.GetNewResourceLocation(newToDo.Id);
-
-            return Created(toDoLocationUrl, newToDo);
+            string toDoLocationUrl = await CreateToDoAsync(toDoItem);
+            return Created(toDoLocationUrl, toDoItem);
         }
         
         [Route("{id}")]
         public async Task<IHttpActionResult> PutToDoAsync(Guid id, [FromBody]ToDo toDoItem)
         {
+            if (!await _retrieveService.IsInDatabaseAsync(id))
+            {
+                string toDoLocationUrl = await CreateToDoAsync(toDoItem);
+                return Created(toDoLocationUrl, toDoItem);
+            }
+
             await _formationService.UpdateToDoAsync(toDoItem);
             return StatusCode(HttpStatusCode.NoContent);
         }
@@ -51,8 +64,23 @@ namespace ToDoList.Api.Controllers
         [Route("{id}")]
         public async Task<IHttpActionResult> DeleteToDoAsync(Guid id)
         {
+            if (!await _retrieveService.IsInDatabaseAsync(id))
+            {
+                return NotFound();
+            }
+
             await _toDoRepository.DeleteToDoAsync(id);
+            _retrieveService.ClearCache();
+
             return StatusCode(HttpStatusCode.NoContent);
+        }
+
+        private async Task<string> CreateToDoAsync(ToDo toDoItem)
+        {
+            toDoItem = await _formationService.CreateToDoAsync(toDoItem);
+            string toDoLocationUrl = _locationService.GetNewResourceLocation(toDoItem.Id);
+
+            return toDoLocationUrl;
         }
     }
 }
